Restructure Topbeat * Use new simplified run methods provided by libbeat * Update libbeat to most recent version * Move brun test into tests folder * Move all topbeat specific code into beat package
