+++ Godeps/Godeps.json
@@ -12,68 +12,55 @@
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/beat",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/cfgfile",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/common",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/logp",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/outputs",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/publisher",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/service",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/cfgfile",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/common",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/logp",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/outputs",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/publisher",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/elastic/libbeat/service",
-			"Comment": "v1.0.0-rc2-16-g7e54c27",
-			"Rev": "7e54c27c23d9c17b343a98edfb8b52cb72a68b1b"
+			"Rev": "007f2e888d457ecf88d606e7a9d36d756f8a500e"
 		},
 		{
 			"ImportPath": "github.com/garyburd/redigo/internal",
