+++ beat/sigar.go
@@ -1,4 +1,4 @@
-package main
+package beat
 
 import (
 	"fmt"
