+++ main.go
@@ -1,10 +1,9 @@
 package main
 
 import (
-	"os"
+	topbeat "github.com/elastic/topbeat/beat"
 
 	"github.com/elastic/libbeat/beat"
-	"github.com/elastic/libbeat/logp"
 )
 
 // You can overwrite these, e.g.: go build -ldflags "-X main.Version 1.0.0-beta3"
@@ -12,19 +11,5 @@ var Version = "1.0.0"
 var Name = "topbeat"
 
 func main() {
-
-	tb := &Topbeat{}
-
-	b := beat.NewBeat(Name, Version, tb)
-
-	b.CommandLineSetup()
-
-	b.LoadConfig()
-	err := tb.Config(b)
-	if err != nil {
-		logp.Critical("Config error: %v", err)
-		os.Exit(1)
-	}
-
-	b.Run()
+	beat.Run(Name, Version, topbeat.New())
 }
