+++ tests/burn.go
